+++ ADOL-C/src/CMakeLists.txt
@@ -14,45 +14,45 @@ target_sources(adolc PRIVATE
                fixpoint.cpp
                fortutils.c
                forward_partx.c
-               fos_forward.c
-               fos_pl_forward.c
+               fos_forward.cpp
+               fos_pl_forward.cpp
                fos_pl_reverse.cpp
-               fos_pl_sig_forward.c
+               fos_pl_sig_forward.cpp
                fos_reverse.cpp
-               fov_forward.c
-               fov_offset_forward.c
-               fov_pl_sig_forward.c
+               fov_forward.cpp
+               fov_offset_forward.cpp
+               fov_pl_sig_forward.cpp
                fos_pl_sig_reverse.cpp
-               fov_pl_forward.c
+               fov_pl_forward.cpp
                fov_reverse.cpp
-               hos_forward.c
+               hos_forward.cpp
                hos_ov_reverse.cpp
                hos_reverse.cpp
-               hov_forward.c
+               hov_forward.cpp
                hov_reverse.cpp
-               hov_wk_forward.c
-               indopro_forward_pl.c
-               indopro_forward_s.c
-               indopro_forward_t.c
-               int_forward_s.c
-               int_forward_t.c
+               hov_wk_forward.cpp
+               indopro_forward_pl.cpp
+               indopro_forward_s.cpp
+               indopro_forward_t.cpp
+               int_forward_s.cpp
+               int_forward_t.cpp
                int_reverse_s.cpp
                int_reverse_t.cpp
                interfaces.cpp
                interfacesf.c
                medipacksupport.cpp
-               nonl_ind_forward_s.c
-               nonl_ind_forward_t.c
-               nonl_ind_old_forward_s.c
-               nonl_ind_old_forward_t.c
+               nonl_ind_forward_s.cpp
+               nonl_ind_forward_t.cpp
+               nonl_ind_old_forward_s.cpp
+               nonl_ind_old_forward_t.cpp
                param.cpp
                revolve.c
                rpl_malloc.c
                storemanager.cpp
                tape_handling.cpp
                taping.c
-               zos_forward.c
-               zos_pl_forward.c
+               zos_forward.cpp
+               zos_pl_forward.cpp
               )
 
 add_subdirectory(drivers)
