+++ ADOL-C/src/fos_forward.cpp
@@ -1,19 +1,19 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     fov_an_forward.c
+ File:     fos_forward.c
  Revision: $Id$
- Contents: fov_an_forward (first-order-vector abs-normal forward mode)
+ Contents: fos_forward (first-order-scalar forward mode)
 
- Copyright (c) Kshitij Kulshrestha
+ Copyright (c) Andrea Walther, Andreas Griewank, Andreas Kowarz,
+               Hristo Mitev, Sebastian Schlenkrich, Jean Utke, Olaf Vogel
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _FOV_ 1
+#define _FOS_ 1
+#define _KEEP_ 1
+#include <uni5_for.cpp>
 #undef _KEEP_
-#define _ABS_NORM_ 1
-#include <uni5_for.c>
-#undef _ABS_NORM_
-#undef _FOV_
+#undef _FOS_
