+++ ADOL-C/src/fov_pl_forward.cpp
@@ -1,18 +1,19 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     hov_forward.c
+ File:     fov_pl_forward.cpp
  Revision: $Id$
- Contents: hov_forward (higher-order-vector forward mode)
+ Contents: fov_an_forward (first-order-vector abs-normal forward mode)
 
- Copyright (c) Andrea Walther, Andreas Griewank, Andreas Kowarz,
-               Hristo Mitev, Sebastian Schlenkrich, Jean Utke, Olaf Vogel
+ Copyright (c) Kshitij Kulshrestha
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _HOV_ 1
+#define _FOV_ 1
 #undef _KEEP_
-#include <uni5_for.c>
-#undef _HOV_
+#define _ABS_NORM_ 1
+#include <uni5_for.cpp>
+#undef _ABS_NORM_
+#undef _FOV_
