+++ ADOL-C/src/fov_pl_sig_forward.cpp
@@ -1,20 +1,19 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     int_forward_s.c
+ File:     fov_an_forward.cpp
  Revision: $Id$
- Contents: int_forward (integer forward mode for bit pattern propagation)
+ Contents: fov_an_forward (first-order-vector abs-normal forward mode)
 
- Copyright (c) Andrea Walther
+ Copyright (c) Kshitij Kulshrestha
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _INDO_ 1
-#define _INDOPRO_ 1
-#define _NTIGHT_ 1
-#include <uni5_for.c>
-#undef _INDO_
-#undef _INDOPRO_
-#undef _NTIGHT_
+#define _FOV_ 1
+#undef _KEEP_
+#define _ABS_NORM_SIG_ 1
+#include <uni5_for.cpp>
+#undef _ABS_NORM_SIG_
+#undef _FOV_
