+++ ADOL-C/src/hov_wk_forward.cpp
@@ -1,19 +1,18 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     hos_forward.c
+ File:     hov_wk_forward.cpp
  Revision: $Id$
- Contents: hos_forward (higher-order-scalar forward mode)
+ Contents: hov_wk_forward (higher-order-vector forward mode with keep)
 
- Copyright (c) Andrea Walther, Andreas Griewank, Andreas Kowarz,
-               Hristo Mitev, Sebastian Schlenkrich, Jean Utke, Olaf Vogel
+ Copyright (c) Andrea Walther
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
-----------------------------------------------------------------------------*/
-#define _HOS_ 1
+ ----------------------------------------------------------------------------*/
+#define _HOV_WK_ 1
 #define _KEEP_ 1
-#include <uni5_for.c>
+#include <uni5_for.cpp>
 #undef _KEEP_
-#undef _HOS_
+#undef _HOV_WK_
