+++ ADOL-C/src/indopro_forward_pl.cpp
@@ -1,18 +1,22 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     int_forward_t.c
+ File:     indopro_forward_pl.cpp
  Revision: $Id$
- Contents: int_forward (integer forward mode tight for bit pattern propagation)
+ Contents: indopro_forward_pl (index domains abs-normal forward mode)
 
- Copyright (c) Andrea Walther, Christo Mitev
+ Copyright (c) Kshitij Kulshrestha
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _INT_FOR_ 1
-#define _TIGHT_ 1
-#include <uni5_for.c>
-#undef _INT_FOR_
-#undef _TIGHT_
+#define _INDO_ 1
+#define _INDOPRO_ 1
+#define _NTIGHT_ 1
+#define _ABS_NORM_ 1
+#include <uni5_for.cpp>
+#undef _ABS_NORM_
+#undef _NTIGHT_
+#undef _INDOPRO_
+#undef _INDO_
