+++ ADOL-C/src/indopro_forward_s.cpp
@@ -1,19 +1,20 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     fos_an_forward.c
+ File:     int_forward_s.cpp
  Revision: $Id$
- Contents: fos_an_forward (first-order-scalar abs-normal forward mode)
+ Contents: int_forward (integer forward mode for bit pattern propagation)
 
- Copyright (c) Kshitij Kulshreshtha
+ Copyright (c) Andrea Walther
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _FOS_ 1
-#undef _KEEP_
-#define _ABS_NORM_ 1
-#include <uni5_for.c>
-#undef _ABS_NORM_
-#undef _FOS_
+#define _INDO_ 1
+#define _INDOPRO_ 1
+#define _NTIGHT_ 1
+#include <uni5_for.cpp>
+#undef _INDO_
+#undef _INDOPRO_
+#undef _NTIGHT_
