+++ ADOL-C/src/int_forward_s.cpp
@@ -1,20 +1,18 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     nonl_ind_forward_t.c
+ File:     int_forward_s.cpp
  Revision: $Id$
- Contents: nonl_ind_forward_t (propagation of nonlinear index domains)
+ Contents: int_forward (integer forward mode safe for bit pattern propagation)
 
- Copyright (c) Andrea Walther
+ Copyright (c) Andrea Walther, Christo Mitev
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _INDO_ 1
-#define _NONLIND_ 1
-#define _TIGHT_ 1
-#include <uni5_for.c>
-#undef _INDO_
-#undef _NONLIND_
-#undef _TIGHT_
+#define _INT_FOR_ 1
+#define _NTIGHT_ 1
+#include <uni5_for.cpp>
+#undef _INT_FOR_
+#undef _NTIGHT_
