+++ ADOL-C/src/int_forward_t.cpp
@@ -1,20 +1,18 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     zos_an_forward.c
+ File:     int_forward_t.cpp
  Revision: $Id$
- Contents: zos_forward (zero-order-scalar abs-normal forward mode)
+ Contents: int_forward (integer forward mode tight for bit pattern propagation)
 
- Copyright (c) Kshitij Kulshreshtha
+ Copyright (c) Andrea Walther, Christo Mitev
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _ZOS_ 1
-#define _KEEP_ 1
-#define _ABS_NORM_ 1
-#include <uni5_for.c>
-#undef _ABS_NORM_
-#undef _KEEP_
-#undef _ZOS_
+#define _INT_FOR_ 1
+#define _TIGHT_ 1
+#include <uni5_for.cpp>
+#undef _INT_FOR_
+#undef _TIGHT_
