+++ ADOL-C/src/nonl_ind_forward_s.cpp
@@ -1,19 +1,20 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     fos_forward.c
+ File:     nonl_ind_forward_s.cpp
  Revision: $Id$
- Contents: fos_forward (first-order-scalar forward mode)
+ Contents: nonl_ind_forward_s (propagation of nonlinear index domains)
 
- Copyright (c) Andrea Walther, Andreas Griewank, Andreas Kowarz,
-               Hristo Mitev, Sebastian Schlenkrich, Jean Utke, Olaf Vogel
+ Copyright (c) Andrea Walther
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _FOS_ 1
-#define _KEEP_ 1
-#include <uni5_for.c>
-#undef _KEEP_
-#undef _FOS_
+#define _INDO_ 1
+#define _NONLIND_ 1
+#define _NTIGHT_ 1
+#include <uni5_for.cpp>
+#undef _INDO_
+#undef _NONLIND_
+#undef _NTIGHT_
