+++ ADOL-C/src/nonl_ind_old_forward_s.cpp
@@ -1,18 +1,22 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     fov_forward.c
+ File:     nonl_ind_forward_s.cpp
  Revision: $Id$
- Contents: fov_forward (first-order-vector forward mode)
+ Contents: nonl_ind_forward_s (propagation of nonlinear index domains)
 
- Copyright (c) Andrea Walther, Andreas Griewank, Andreas Kowarz,
-               Hristo Mitev, Sebastian Schlenkrich, Jean Utke, Olaf Vogel
+ Copyright (c) Andrea Walther, Kshitij Kulshreshtha
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _FOV_ 1
-#undef _KEEP_
-#include <uni5_for.c>
-#undef _FOV_
+#define _INDO_ 1
+#define _NONLIND_OLD_ 1
+#define _INDOPRO_ 1
+#define _NTIGHT_ 1
+#include <uni5_for.cpp>
+#undef _INDO_
+#undef _NONLIND_OLD_
+#undef _INDOPRO_
+#undef _NTIGHT_
