+++ ADOL-C/src/nonl_ind_old_forward_t.cpp
@@ -1,10 +1,10 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     int_forward_t.c
+ File:     nonl_ind_forward_t.cpp
  Revision: $Id$
- Contents: int_forward (integer forward mode for bit pattern propagation)
+ Contents: nonl_ind_forward_t (propagation of nonlinear index domains)
 
- Copyright (c) Andrea Walther
+ Copyright (c) Andrea Walther, Kshitij Kulshreshtha
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
@@ -12,9 +12,11 @@
 
 ----------------------------------------------------------------------------*/
 #define _INDO_ 1
+#define _NONLIND_OLD_ 1
 #define _INDOPRO_ 1
 #define _TIGHT_ 1
-#include <uni5_for.c>
+#include <uni5_for.cpp>
 #undef _INDO_
+#undef _NONLIND_OLD_
 #undef _INDOPRO_
 #undef _TIGHT_
