+++ ADOL-C/src/zos_pl_forward.cpp
@@ -1,22 +1,20 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     nonl_ind_forward_t.c
+ File:     zos_an_forward.cpp
  Revision: $Id$
- Contents: nonl_ind_forward_t (propagation of nonlinear index domains)
+ Contents: zos_forward (zero-order-scalar abs-normal forward mode)
 
- Copyright (c) Andrea Walther, Kshitij Kulshreshtha
+ Copyright (c) Kshitij Kulshreshtha
 
  This file is part of ADOL-C. This software is provided as open source.
  Any use, reproduction, or distribution of the software constitutes
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#define _INDO_ 1
-#define _NONLIND_OLD_ 1
-#define _INDOPRO_ 1
-#define _TIGHT_ 1
-#include <uni5_for.c>
-#undef _INDO_
-#undef _NONLIND_OLD_
-#undef _INDOPRO_
-#undef _TIGHT_
+#define _ZOS_ 1
+#define _KEEP_ 1
+#define _ABS_NORM_ 1
+#include <uni5_for.cpp>
+#undef _ABS_NORM_
+#undef _KEEP_
+#undef _ZOS_
